% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAgibbs.R
\name{CAgibbs}
\alias{CAgibbs}
\title{Gibbs Sampler}
\usage{
CAgibbs(obs_network, hired, p_prior, q_prior, d = 0.15, burn = 1000,
  thin = 25, numpostdraws = 1000)
}
\arguments{
\item{obs_network}{The adjacency matrix A of the observed directed network. A is an n by n matrix, where n is the number of nodes. A_ij equals 1 if a tie exists between i and j. A_ij equals 0 if there is no tie between i and j. A_ii equals 0 by default. If fitting collaboration models on multiple time points for the same set of people, n should be the full number of people. Values of \code{obs_network} for people who have not been hired by that time will be 0.}

\item{hired}{An indicator vector of length n indicating whether each person in \code{obs_network} had been hired by this point in time.}

\item{p_prior}{Vector of hyperparameters for the prior Beta distribution of p. Specify both alpha and beta for the Beta(alpha,beta) distribution, e.g. \code{c(2,8)}.}

\item{q_prior}{Vector of hyperparameters for the prior Beta distribution of q. Specify both alpha and beta for the Beta(alpha,beta) distribution, e.g. \code{c(2,8)}.}

\item{d}{The density used for Theta_ij if i and j both hired by the observed time. If i and/or j not hired, Theta_ij is zero. Defaults to 0.15, a reasonable network density.}

\item{burn}{Burn-in period for Gibbs sampler. Defaults to 1000 iterations.}

\item{thin}{Thinning interval for Gibbs sampler. Defaults to 25 iterations.}

\item{numpostdraws}{Number of postrior draws to keep. Sampler will run \code{numpostdraws*thin + burn} iterations in total, and keep just nonpostdraws. Defaults to 1000 iterations.}
}
\value{
A list consisting of
  \item{p}{Matrix (dimension numpostdraws x n) where each row is vector of p_i's for a posterior draw.}
  \item{q}{Matrix (dimension numpostdraws x n) where each row is vector of q_i's for a posterior draw.}
  \item{Theta_prop}{Proportion of times each tie appears in latent adjacency matrices over all draws. The sum of \code{Theta_all} divided by \code{numpostdraws}.}
  \item{Theta_all}{List of all latent adjacency matrices.}
  \item{dens_vec}{Vector consisting of the network density for each posterior draw.}
}
\description{
Gibbs Sampler
}
\examples{
  # This example uses data from Year 1
  data(smYr1)
  data(smYr1_hire)
  year1 <- CAgibbs(smYr1, smYr1_hire, c(2,8), c(2,8), burn=100, thin=5, numpostdraws=100)
    # in practice run for longer

  colMeans(year1$p_mat) #posterior mean for p (false negative) for each person at time 1
  colMeans(year1$q_mat) #posterior mean for q (false positive) for each person at time 1

  threshplot(year1$Theta_prop) #latent network only
  threshplot(year1$Theta_prop, obs=smYr1) #latent network side by side with observed network
}
