% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smYr1}
\alias{smYr1}
\title{Education collaboration ties, year 1.}
\format{A matrix with 37 rows and 37 columns}
\usage{
smYr1
}
\description{
Adjacency matrix showing education collaboration ties, year 1 (2003-2004 academic year).
}
\keyword{datasets}
