% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smYr1_hire}
\alias{smYr1_hire}
\title{Hire list, year 1.}
\format{A vector with 37 entries.}
\usage{
smYr1_hire
}
\description{
Indicator vector showing whether or not each person had been hired by year 1 (2003-2004 academic year).
}
\keyword{datasets}
