% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshplot.R
\name{threshplot}
\alias{threshplot}
\title{Threshold plots for posterior latent networks}
\usage{
threshplot(w_adj, thresh = 0.1, obs = NULL)
}
\arguments{
\item{w_adj}{Weighted adjacency matrix for latent network (symmetric).}

\item{thresh}{Threshold (betweeen 0 and 1) for an edge to be plotted. Defaults to 0.1 (i.e., edges appearing in at least 10\% of posterior latent networks will be shown).}

\item{obs}{The adjacency matric of the observed network on the same vertices as \code{w_adj}. Default is NULL. If provided, plots the latent and observed networks side by side.}
}
\value{
A threshold plot
}
\description{
Threshold plots for posterior latent networks
}
\examples{
  data(smYr1)
  data(smYr1_hire)
  year1 <- CAgibbs(smYr1, smYr1_hire, c(2,8), c(2,8), burn=100, thin=5, numpostdraws=100)
    # in practice run for longer

  threshplot(year1$Theta_prop) #latent network only
  threshplot(year1$Theta_prop, obs=smYr1) #latent network side by side with observed network
}
